zsh: disable autocd option.
Despite the documentation saying that it only autocd's when no command
of that directory name is possible, running `tudu` in a directory
containing `tudu` resulted in `cd tudu`, not sure what's going on.
Anyway, it's probably not a bad idea to get back into the habit of using
`cd` explicitly anyway.